refactor: merge companion files (continue->outline, foreshadow_consistency->foreshadow, agent_i18n->agent)
