--- internal/agent/agent_i18n.go
@@ -1,39 +1,0 @@
-package agent
-
-import (
-	"fmt"
-	"showmethestory/internal/i18n"
-)
-
-// ponytail: per-tool-call scratch fields on AgentContext; single-threaded agent loop only.
-func (ctx *AgentContext) clearToolMsg() {
-	ctx.toolMsgKey = ""
-	ctx.toolMsgArgs = nil
-}
-
-func (ctx *AgentContext) setToolMsg(key string, args ...any) {
-	ctx.toolMsgKey = key
-	ctx.toolMsgArgs = i18n.MsgArgs(args...)
-}
-
-func (ctx *AgentContext) takeToolMsg() (string, []string) {
-	k, a := ctx.toolMsgKey, ctx.toolMsgArgs
-	ctx.clearToolMsg()
-	return k, a
-}
-
-func projectLang(ctx *AgentContext) string {
-	if ctx == nil || ctx.Config == nil {
-		return i18n.LangZH
-	}
-	return i18n.NormalizeLanguage(ctx.Config.Language)
-}
-
-func agentMsg(ctx *AgentContext, key string, args ...any) string {
-	ctx.setToolMsg(key, args...)
-	return i18n.T(projectLang(ctx), key, args...)
-}
-
-func agentErr(ctx *AgentContext, key string, args ...any) error {
-	return fmt.Errorf("%s", i18n.T(projectLang(ctx), key, args...))
-}
